Create a context manager for fcntl.flock with LOCK_SH
To support the capability of the backup system to copy a target file to
a temporary location, it would be nice to put a shared lock on the file;
even if it is only advisory, it _might_ convince another process to wait
while the copy is made.  It will also at least give warning in the case
the file couldn't be locked.