Reduce calls to PutObjectRetention
If local data indicates the object's retention is long enough, there is
no need to incur the expense of extending the retention.  Even if local
data does not so indicate, because of "space folding", another system
might have uploaded this content, and asking for the object retention
information is an order of magnitude cheaper than extending it if not
needed.  This will cost _slightly_ more for single system users that
never see folding within their own selected files.