Allow the backup auditor role to browse in AWS S3 console
To view the bucket in the AWS console, the auditor needs to be able to
list the bucket with an empty string `prefix` and the slash delimiter.

The auditor also needs to be able to verify the existence of the correct
snapshots and to verify retention settings of any object in the bucket.

Closes #5